Split configuration form
refs #54
