+++ application/controllers/ConfigController.php
@@ -2,7 +2,8 @@
 
 namespace Icinga\Module\Graphite\Controllers;
 
-use Icinga\Module\Graphite\Forms\ConfigForm;
+use Icinga\Module\Graphite\Forms\Config\AdvancedForm;
+use Icinga\Module\Graphite\Forms\Config\BackendForm;
 use Icinga\Web\Controller;
 
 class ConfigController extends Controller
@@ -13,10 +14,17 @@ class ConfigController extends Controller
         parent::init();
     }
 
-    public function indexAction()
+    public function backendAction()
     {
-        $this->view->form = $form = new ConfigForm();
+        $this->view->form = $form = new BackendForm();
         $form->setIniConfig($this->Config())->handleRequest();
         $this->view->tabs = $this->Module()->getConfigTabs()->activate('backend');
+    }
+
+    public function advancedAction()
+    {
+        $this->view->form = $form = new AdvancedForm();
+        $form->setIniConfig($this->Config())->handleRequest();
+        $this->view->tabs = $this->Module()->getConfigTabs()->activate('advanced');
     }
 }
