+++ application/views/scripts/config/advanced.phtml
@@ -1,0 +1,7 @@
+<div class="controls">
+    <?= /** @var \Icinga\Web\Widget\Tabs $tabs */ $tabs ?>
+</div>
+
+<div class="content">
+    <?= /** @var \Icinga\Module\Graphite\Forms\Config\AdvancedForm $form */ $form ?>
+</div>
