+++ application/views/scripts/config/backend.phtml
@@ -3,5 +3,5 @@
 </div>
 
 <div class="content">
-    <?= /** @var \Icinga\Module\Graphite\Forms\ConfigForm $form */ $form ?>
+    <?= /** @var \Icinga\Module\Graphite\Forms\Config\BackendForm $form */ $form ?>
 </div>
